Moved more functions into the preview module based on comments from Monica.
